requirements.txt
# ============================================
# Best Time Post API - Requirements
# ============================================

# --- API Framework ---
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.5.0

# --- Rate Limiting & Caching ---
slowapi==0.1.9
cachetools==5.3.2

# --- Apify & Data Collection ---
apify-client==1.6.0
requests==2.31.0

# --- Database ---
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# --- Data Processing ---
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
joblib==1.3.2

# --- ONNX Inference ---
onnx==1.15.0
onnxruntime==1.16.3

# --- Deep Learning (for export script) ---
torch==2.1.2

# --- Logging ---
loguru==0.7.2

# --- Time & Date ---
pytz==2023.3.post1
python-dateutil==2.8.2

# --- Machine Learning Visualization & Boosting ---
lightgbm==4.1.0
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# --- Time Series Forecasting ---
neuralprophet==0.7.0
onnx==1.15.0
onnxruntime==1.17.0


# --- Machine Learning Visualization & Boosting ---
lightgbm==4.1.0
matplotlib==3.8.2
seaborn==0.13.0
